moar warning fixes: unused parameters, and in StringCollection changed the type of current from int to size_t
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@2904 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/include/tulip/StringCollection.h
@@ -32,7 +32,7 @@ namespace tlp {
 struct TLP_SCOPE StringCollection : public std::vector<std::string> {
 
 private:
-  int current;
+  size_t current;
 
 public:
   StringCollection();
